requirements.txt
streamlit
forex-python
qrcode
pillow
pyzbar
pandas